requirements/base.txt
asgiref==3.8.1
async-timeout==4.0.3
Django==5.0.4
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.15.1
psycopg2==2.9.9
python-dotenv==1.0.1
redis==5.0.4
sqlparse==0.5.0
typing_extensions==4.11.0
